Fixed a bug when generating player_results.json
The keys for player id in player_results.json would sometimes be set as the player's full name and not the player id, due to errors on Liquipedia for 40+ players.

Therefore changed liquipedia query from "has ids" to "has id", and made sure that player entity is actually equal to the player id.